+++ lib/knapsack/distributors/base_distributor.rb
@@ -1,11 +1,11 @@
 module Knapsack
   module Distributors
     class BaseDistributor
-      attr_reader :report, :node_specs, :spec_pattern
+      attr_reader :report, :node_tests, :test_file_pattern
 
       def initialize(args={})
         @report = args[:report] || raise('Missing report')
-        @spec_pattern = args[:spec_pattern] || raise('Missing spec_pattern')
+        @test_file_pattern = args[:test_file_pattern] || raise('Missing test_file_pattern')
         @ci_node_total = args[:ci_node_total] || raise('Missing ci_node_total')
         @ci_node_index = args[:ci_node_index] || raise('Missing ci_node_index')
       end
@@ -18,35 +18,35 @@ module Knapsack
         @ci_node_index.to_i
       end
 
-      def specs_for_current_node
-        specs_for_node(ci_node_index)
+      def tests_for_current_node
+        tests_for_node(ci_node_index)
       end
 
-      def specs_for_node(node_index)
-        assign_spec_files_to_node
-        post_specs_for_node(node_index)
+      def tests_for_node(node_index)
+        assign_test_files_to_node
+        post_tests_for_node(node_index)
       end
 
-      def assign_spec_files_to_node
-        default_node_specs
-        post_assign_spec_files_to_node
+      def assign_test_files_to_node
+        default_node_tests
+        post_assign_test_files_to_node
       end
 
-      def all_specs
-        @all_specs ||= Dir[spec_pattern]
+      def all_tests
+        @all_tests ||= Dir[test_file_pattern]
       end
 
       protected
 
-      def post_specs_for_node(node_index)
+      def post_tests_for_node(node_index)
         raise NotImplementedError
       end
 
-      def post_assign_spec_files_to_node
+      def post_assign_test_files_to_node
         raise NotImplementedError
       end
 
-      def default_node_specs
+      def default_node_tests
         raise NotImplementedError
       end
     end