+++ spec/knapsack/allocator_builder_spec.rb
@@ -6,19 +6,19 @@ describe Knapsack::AllocatorBuilder do
   let(:knapsack_report) { instance_double(Knapsack::Report) }
 
   let(:adapter_report_path) { adapter_class::REPORT_PATH }
-  let(:adapter_spec_pattern) { adapter_class::TEST_DIR_PATTERN }
+  let(:adapter_test_file_pattern) { adapter_class::TEST_DIR_PATTERN }
 
   let(:env_ci_node_total) { double }
   let(:env_ci_node_index) { double }
   let(:env_report_path) { nil }
-  let(:env_spec_pattern) { nil }
+  let(:env_test_file_pattern) { nil }
 
   describe '#allocator' do
     subject { allocator_builder.allocator }
 
     before do
       expect(Knapsack::Config::Env).to receive(:report_path).and_return(env_report_path)
-      expect(Knapsack::Config::Env).to receive(:spec_pattern).and_return(env_spec_pattern)
+      expect(Knapsack::Config::Env).to receive(:test_file_pattern).and_return(env_test_file_pattern)
       expect(Knapsack::Config::Env).to receive(:ci_node_total).and_return(env_ci_node_total)
       expect(Knapsack::Config::Env).to receive(:ci_node_index).and_return(env_ci_node_index)
 
@@ -35,7 +35,7 @@ describe Knapsack::AllocatorBuilder do
         let(:allocator_args) do
           {
             report: report,
-            spec_pattern: adapter_spec_pattern,
+            test_file_pattern: adapter_test_file_pattern,
             ci_node_total: env_ci_node_total,
             ci_node_index: env_ci_node_index
           }
@@ -50,7 +50,7 @@ describe Knapsack::AllocatorBuilder do
         let(:allocator_args) do
           {
             report: report,
-            spec_pattern: adapter_spec_pattern,
+            test_file_pattern: adapter_test_file_pattern,
             ci_node_total: env_ci_node_total,
             ci_node_index: env_ci_node_index
           }
@@ -59,13 +59,13 @@ describe Knapsack::AllocatorBuilder do
         it { should eql allocator }
       end
 
-      context 'when ENV spec_pattern has value' do
-        let(:env_spec_pattern) { 'custom_spec/**/*_spec.rb' }
+      context 'when ENV test_file_pattern has value' do
+        let(:env_test_file_pattern) { 'custom_spec/**/*_spec.rb' }
         let(:report_config) { { report_path: adapter_report_path } }
         let(:allocator_args) do
           {
             report: report,
-            spec_pattern: env_spec_pattern,
+            test_file_pattern: env_test_file_pattern,
             ci_node_total: env_ci_node_total,
             ci_node_index: env_ci_node_index
           }