+++ spec/knapsack/allocator_spec.rb
@@ -1,56 +1,56 @@
 describe Knapsack::Allocator do
-  let(:spec_pattern) { nil }
+  let(:test_file_pattern) { nil }
   let(:args) do
     {
       ci_node_total: nil,
       ci_node_index: nil,
-      spec_pattern: spec_pattern,
+      test_file_pattern: test_file_pattern,
       report: nil
     }
   end
   let(:report_distributor) { instance_double(Knapsack::Distributors::ReportDistributor) }
   let(:leftover_distributor) { instance_double(Knapsack::Distributors::LeftoverDistributor) }
-  let(:report_specs) { ['a_spec.rb', 'b_spec.rb'] }
-  let(:leftover_specs) { ['c_spec.rb', 'd_spec.rb'] }
-  let(:node_specs) { report_specs + leftover_specs }
+  let(:report_tests) { ['a_spec.rb', 'b_spec.rb'] }
+  let(:leftover_tests) { ['c_spec.rb', 'd_spec.rb'] }
+  let(:node_tests) { report_tests + leftover_tests }
   let(:allocator) { described_class.new(args) }
 
   before do
     expect(Knapsack::Distributors::ReportDistributor).to receive(:new).with(args).and_return(report_distributor)
     expect(Knapsack::Distributors::LeftoverDistributor).to receive(:new).with(args).and_return(leftover_distributor)
-    allow(report_distributor).to receive(:specs_for_current_node).and_return(report_specs)
-    allow(leftover_distributor).to receive(:specs_for_current_node).and_return(leftover_specs)
+    allow(report_distributor).to receive(:tests_for_current_node).and_return(report_tests)
+    allow(leftover_distributor).to receive(:tests_for_current_node).and_return(leftover_tests)
   end
 
-  describe '#report_node_specs' do
-    subject { allocator.report_node_specs }
-    it { should eql report_specs }
+  describe '#report_node_tests' do
+    subject { allocator.report_node_tests }
+    it { should eql report_tests }
   end
 
-  describe '#leftover_node_specs' do
-    subject { allocator.leftover_node_specs }
-    it { should eql leftover_specs }
+  describe '#leftover_node_tests' do
+    subject { allocator.leftover_node_tests }
+    it { should eql leftover_tests }
   end
 
-  describe '#node_specs' do
-    subject { allocator.node_specs }
-    it { should eql node_specs }
+  describe '#node_tests' do
+    subject { allocator.node_tests }
+    it { should eql node_tests }
   end
 
-  describe '#stringify_node_specs' do
-    subject { allocator.stringify_node_specs }
-    it { should eql node_specs.join(' ') }
+  describe '#stringify_node_tests' do
+    subject { allocator.stringify_node_tests }
+    it { should eql node_tests.join(' ') }
   end
 
-  describe '#spec_dir' do
-    let(:spec_pattern) { "spec_dir/**/*_spec.rb" }
+  describe '#test_dir' do
+    let(:test_file_pattern) { "test_dir/**/*_spec.rb" }
 
-    subject { allocator.spec_dir }
+    subject { allocator.test_dir }
 
     before do
-      expect(report_distributor).to receive(:spec_pattern).and_return(spec_pattern)
+      expect(report_distributor).to receive(:test_file_pattern).and_return(test_file_pattern)
     end
 
-    it { should eql 'spec_dir/' }
+    it { should eql 'test_dir/' }
   end
 end