+++ spec/knapsack/config/env_spec.rb
@@ -63,13 +63,13 @@ describe Knapsack::Config::Env do
     end
   end
 
-  describe '.spec_pattern' do
-    subject { described_class.spec_pattern }
+  describe '.test_file_pattern' do
+    subject { described_class.test_file_pattern }
 
     context 'when ENV exists' do
-      let(:spec_pattern) { 'custom_spec/**/*_spec.rb' }
-      before { stub_const("ENV", { 'KNAPSACK_SPEC_PATTERN' => spec_pattern }) }
-      it { should eql spec_pattern }
+      let(:test_file_pattern) { 'custom_spec/**/*_spec.rb' }
+      before { stub_const("ENV", { 'KNAPSACK_TEST_FILE_PATTERN' => test_file_pattern }) }
+      it { should eql test_file_pattern }
     end
 
     context "when ENV doesn't exist" do