+++ spec/knapsack/distributors/leftover_distributor_spec.rb
@@ -7,11 +7,11 @@ describe Knapsack::Distributors::LeftoverDistributor do
       'd_spec.rb' => 2.5,
     }
   end
-  let(:spec_pattern) { 'spec/**/*_spec.rb' }
+  let(:test_file_pattern) { 'spec/**/*_spec.rb' }
   let(:default_args) do
     {
       report: report,
-      spec_pattern: spec_pattern,
+      test_file_pattern: test_file_pattern,
       ci_node_total: '1',
       ci_node_index: '0'
     }
@@ -20,15 +20,15 @@ describe Knapsack::Distributors::LeftoverDistributor do
   let(:custom_args) { {} }
   let(:distributor) { described_class.new(args) }
 
-  describe '#report_specs' do
-    subject { distributor.report_specs }
+  describe '#report_tests' do
+    subject { distributor.report_tests }
     it { should eql ['a_spec.rb', 'b_spec.rb', 'c_spec.rb', 'd_spec.rb'] }
   end
 
-  describe '#all_specs' do
-    subject { distributor.all_specs }
+  describe '#all_tests' do
+    subject { distributor.all_tests }
 
-    context 'when given spec pattern' do
+    context 'when given test_file_pattern' do
       context 'spec/**/*_spec.rb' do
         it { should_not be_empty }
         it { should include 'spec/knapsack/tracker_spec.rb' }
@@ -36,7 +36,7 @@ describe Knapsack::Distributors::LeftoverDistributor do
       end
 
       context 'spec_examples/**/*_spec.rb' do
-        let(:spec_pattern) { 'spec_examples/**/*_spec.rb' }
+        let(:test_file_pattern) { 'spec_examples/**/*_spec.rb' }
 
         it { should_not be_empty }
         it { should include 'spec_examples/fast/1_spec.rb' }
@@ -44,22 +44,22 @@ describe Knapsack::Distributors::LeftoverDistributor do
       end
     end
 
-    context 'when fake spec pattern' do
-      let(:spec_pattern) { 'fake_pattern' }
+    context 'when fake test_file_pattern' do
+      let(:test_file_pattern) { 'fake_pattern' }
       it { should be_empty }
     end
 
-    context 'when missing spec pattern' do
-      let(:spec_pattern) { nil }
-      it { expect { subject }.to raise_error('Missing spec_pattern') }
+    context 'when missing test_file_pattern' do
+      let(:test_file_pattern) { nil }
+      it { expect { subject }.to raise_error('Missing test_file_pattern') }
     end
   end
 
-  describe '#leftover_specs' do
-    subject { distributor.leftover_specs }
+  describe '#leftover_tests' do
+    subject { distributor.leftover_tests }
 
     before do
-      expect(distributor).to receive(:all_specs).and_return([
+      expect(distributor).to receive(:all_tests).and_return([
         'a_spec.rb',
         'b_spec.rb',
         'c_spec.rb',
@@ -74,7 +74,7 @@ describe Knapsack::Distributors::LeftoverDistributor do
 
   context do
     let(:custom_args) { { ci_node_total: 3 } }
-    let(:leftover_specs) {[
+    let(:leftover_tests) {[
       'a_spec.rb',
       'b_spec.rb',
       'c_spec.rb',
@@ -85,16 +85,16 @@ describe Knapsack::Distributors::LeftoverDistributor do
     ]}
 
     before do
-      expect(distributor).to receive(:leftover_specs).and_return(leftover_specs)
+      expect(distributor).to receive(:leftover_tests).and_return(leftover_tests)
     end
 
-    describe '#assign_spec_files_to_node' do
+    describe '#assign_test_files_to_node' do
       before do
-        distributor.assign_spec_files_to_node
+        distributor.assign_test_files_to_node
       end
 
       it do
-        expect(distributor.node_specs[0]).to eql([
+        expect(distributor.node_tests[0]).to eql([
           'a_spec.rb',
           'd_spec.rb',
           'g_spec.rb',
@@ -102,23 +102,23 @@ describe Knapsack::Distributors::LeftoverDistributor do
       end
 
       it do
-        expect(distributor.node_specs[1]).to eql([
+        expect(distributor.node_tests[1]).to eql([
           'b_spec.rb',
           'e_spec.rb',
         ])
       end
 
       it do
-        expect(distributor.node_specs[2]).to eql([
+        expect(distributor.node_tests[2]).to eql([
           'c_spec.rb',
           'f_spec.rb',
         ])
       end
     end
 
-    describe '#specs_for_node' do
+    describe '#tests_for_node' do
       it do
-        expect(distributor.specs_for_node(1)).to eql([
+        expect(distributor.tests_for_node(1)).to eql([
           'b_spec.rb',
           'e_spec.rb',
         ])